+++ vcd/datasource_vcd_lb_app_rule.go
@@ -11,16 +11,17 @@ func datasourceVcdLBAppRule() *schema.Resource {
 		Read: datasourceVcdLBAppRuleRead,
 		Schema: map[string]*schema.Schema{
 			"org": {
-				Type:        schema.TypeString,
-				Optional:    true,
-				ForceNew:    true,
-				Description: "vCD organization in which the LB Application Rule is located",
+				Type:     schema.TypeString,
+				Optional: true,
+				ForceNew: true,
+				Description: "The name of organization to use, optional if defined at provider " +
+					"level. Useful when connected as sysadmin working across different organizations",
 			},
 			"vdc": {
 				Type:        schema.TypeString,
 				Optional:    true,
 				ForceNew:    true,
-				Description: "vCD virtual datacenter in which the LB Application Rule is located",
+				Description: "The name of VDC to use, optional if defined at provider level",
 			},
 			"edge_gateway": &schema.Schema{
 				Type:        schema.TypeString,