+++ vcd/provider.go
@@ -109,6 +109,8 @@ func Provider() terraform.ResourceProvider {
 			"vcd_lb_app_profile":     resourceVcdLBAppProfile(),     // 2.4
 			"vcd_lb_app_rule":        resourceVcdLBAppRule(),        // 2.4
 			"vcd_lb_virtual_server":  resourceVcdLBVirtualServer(),  // 2.4
+			"vcd_nsxv_dnat":          resourceVcdNsxvDnat(),         // 2.5
+			"vcd_nsxv_snat":          resourceVcdNsxvSnat(),         // 2.5
 		},
 
 		DataSourcesMap: map[string]*schema.Resource{
@@ -122,6 +124,8 @@ func Provider() terraform.ResourceProvider {
 			"vcd_lb_app_profile":     datasourceVcdLBAppProfile(),     // 2.4
 			"vcd_lb_app_rule":        datasourceVcdLBAppRule(),        // 2.4
 			"vcd_lb_virtual_server":  datasourceVcdLbVirtualServer(),  // 2.4
+			"vcd_nsxv_dnat":          datasourceVcdNsxvDnat(),         // 2.5
+			"vcd_nsxv_snat":          datasourceVcdNsxvSnat(),         // 2.5
 		},
 
 		ConfigureFunc: providerConfigure,