+++ vcd/resource_vcd_catalog_media.go
@@ -23,6 +23,8 @@ func resourceVcdCatalogMedia() *schema.Resource {
 				Type:     schema.TypeString,
 				Optional: true,
 				ForceNew: true,
+				Description: "The name of organization to use, optional if defined at provider " +
+					"level. Useful when connected as sysadmin working across different organizations",
 			},
 			"catalog": &schema.Schema{
 				Type:        schema.TypeString,