+++ vcd/resource_vcd_vapp.go
@@ -28,12 +28,15 @@ func resourceVcdVApp() *schema.Resource {
 				Required: false,
 				Optional: true,
 				ForceNew: true,
+				Description: "The name of organization to use, optional if defined at provider " +
+					"level. Useful when connected as sysadmin working across different organizations",
 			},
 			"vdc": {
-				Type:     schema.TypeString,
-				Required: false,
-				Optional: true,
-				ForceNew: true,
+				Type:        schema.TypeString,
+				Required:    false,
+				Optional:    true,
+				ForceNew:    true,
+				Description: "The name of VDC to use, optional if defined at provider level",
 			},
 			"template_name": {
 				Type:     schema.TypeString,